requirements.txt
py-pure-client
pyyaml
